Consolidate React into the keyboardist/react subpath
One package instead of two lockstep ones: React hooks and components
now ship as `import ... from "keyboardist/react"`, and the separate
react-keyboardist package is retired (its npm name stays frozen at
0.2.0; v3 of it was never published).

- exports map gains "./react" (types-first, ESM)
- react is an optional peerDependency — vanilla users see no warnings
- tsdown builds both entries with a shared chunk: core stays
  directive-free, dist/react/index.js starts with "use client"
  (both asserted in check:publish; attw validates both subpaths)
- src/react/* and test/react/* moved from packages/react-keyboardist,
  package imports rewritten to relative; all 96 tests green (67 core +
  29 react) in the merged package
- changesets: fixed group removed, react-keyboardist entries rewritten
  as keyboardist entries; only keyboardist bumps (major -> 3.0.0)
- website imports keyboardist/react; docs merged into the single README
