+++ .changeset/config.json
@@ -5,7 +5,7 @@
     { "repo": "soska/keyboardist" }
   ],
   "commit": false,
-  "fixed": [["keyboardist", "react-keyboardist"]],
+  "fixed": [],
   "linked": [],
   "access": "public",
   "baseBranch": "master",
