+++ apps/website/index.html
@@ -72,8 +72,8 @@
         <h2>React</h2>
         <div class="instructions">
           This section is a React island using
-          <a href="https://www.npmjs.com/package/react-keyboardist"
-            >react-keyboardist</a
+          <a href="https://www.npmjs.com/package/keyboardist"
+            >keyboardist/react</a
           >
           — hooks and components over the same shared listener as the vanilla
           demos above.
