+++ apps/website/vite.config.ts
@@ -5,18 +5,24 @@ import { defineConfig } from "vite";
 export default defineConfig({
   plugins: [react()],
   resolve: {
-    alias: {
-      // Resolve the workspace packages to their source so the demo
-      // hot-reloads library changes without a build step.
-      "react-keyboardist": fileURLToPath(
-        new URL(
-          "../../packages/react-keyboardist/src/index.ts",
-          import.meta.url,
+    // Resolve the workspace package to its source so the demo hot-reloads
+    // library changes without a build step. Most-specific alias first.
+    alias: [
+      {
+        find: "keyboardist/react",
+        replacement: fileURLToPath(
+          new URL(
+            "../../packages/keyboardist/src/react/index.ts",
+            import.meta.url,
+          ),
         ),
-      ),
-      keyboardist: fileURLToPath(
-        new URL("../../packages/keyboardist/src/index.ts", import.meta.url),
-      ),
-    },
+      },
+      {
+        find: "keyboardist",
+        replacement: fileURLToPath(
+          new URL("../../packages/keyboardist/src/index.ts", import.meta.url),
+        ),
+      },
+    ],
   },
 });
