+++ packages/keyboardist/package.json
@@ -1,7 +1,7 @@
 {
   "name": "keyboardist",
   "version": "2.0.2",
-  "description": "A declarative way to add keyboard shortcuts to your browser applications",
+  "description": "A declarative way to add keyboard shortcuts to your browser applications — with React hooks and components at keyboardist/react",
   "author": "Armando Sosa <arm.sosa@gmail.com>",
   "license": "MIT",
   "type": "module",
@@ -10,6 +10,10 @@
     ".": {
       "types": "./dist/index.d.ts",
       "default": "./dist/index.js"
+    },
+    "./react": {
+      "types": "./dist/react/index.d.ts",
+      "default": "./dist/react/index.js"
     }
   },
   "types": "./dist/index.d.ts",
@@ -20,7 +24,9 @@
     "keyboard",
     "shortcuts",
     "hotkeys",
-    "keybindings"
+    "keybindings",
+    "react",
+    "hooks"
   ],
   "homepage": "https://github.com/soska/keyboardist",
   "repository": {
@@ -36,19 +42,32 @@
     "access": "public",
     "provenance": true
   },
+  "peerDependencies": {
+    "react": "^18.0.0 || ^19.0.0"
+  },
+  "peerDependenciesMeta": {
+    "react": {
+      "optional": true
+    }
+  },
   "scripts": {
     "build": "tsdown",
     "dev": "tsdown --watch",
     "test": "vitest run",
     "test:watch": "vitest",
     "typecheck": "tsc --noEmit",
-    "check:publish": "publint && attw --pack . --profile esm-only"
+    "check:publish": "publint && attw --pack . --profile esm-only && node --input-type=module -e \"import{readFileSync}from'node:fs';const react=readFileSync('dist/react/index.js','utf8');const core=readFileSync('dist/index.js','utf8');if(!react.startsWith('\\\"use client\\\"')&&!react.startsWith(\\\"'use client'\\\")){console.error('dist/react missing use client directive');process.exit(1)};if(core.includes('use client')){console.error('core entry must not carry use client');process.exit(1)}\""
   },
   "devDependencies": {
     "@arethetypeswrong/cli": "^0.18.5",
     "@testing-library/dom": "^10.4.1",
+    "@testing-library/react": "^16.3.2",
+    "@types/react": "^19.2.17",
+    "@types/react-dom": "^19.2.3",
     "happy-dom": "^20.11.1",
     "publint": "^0.3.21",
+    "react": "^19.2.8",
+    "react-dom": "^19.2.8",
     "tsdown": "^0.22.13",
     "typescript": "^7.0.2",
     "vitest": "^4.1.10"
