+++ packages/keyboardist/tsconfig.json
@@ -1,4 +1,7 @@
 {
   "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "jsx": "react-jsx"
+  },
   "include": ["src", "test", "tsdown.config.ts", "vitest.config.ts"]
 }
