+++ packages/keyboardist/tsdown.config.ts
@@ -1,7 +1,10 @@
 import { defineConfig } from "tsdown";
 
 export default defineConfig({
-  entry: ["src/index.ts"],
+  entry: {
+    index: "src/index.ts",
+    "react/index": "src/react/index.ts",
+  },
   format: ["esm"],
   platform: "browser",
   dts: true,
