+++ packages/keyboardist/vitest.config.ts
@@ -3,5 +3,6 @@ import { defineConfig } from "vitest/config";
 export default defineConfig({
   test: {
     environment: "happy-dom",
+    setupFiles: ["./test/react/setup.ts"],
   },
 });
