--- packages/react-keyboardist/tsconfig.build.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    // No paths override here: the dts build must treat `keyboardist` as an
-    // external package (its published types), not follow the source mapping —
-    // otherwise tsc emits stray .d.ts files into ../keyboardist/src.
-    "paths": {}
-  },
-  "include": ["src"]
-}
