--- packages/react-keyboardist/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "jsx": "react-jsx",
-    "paths": {
-      "keyboardist": ["../keyboardist/src/index.ts"]
-    }
-  },
-  "include": ["src", "test", "tsdown.config.ts", "vitest.config.ts"]
-}
