--- packages/react-keyboardist/tsdown.config.ts
@@ -1,10 +1,0 @@
-import { defineConfig } from "tsdown";
-
-export default defineConfig({
-  entry: ["src/index.ts"],
-  format: ["esm"],
-  platform: "browser",
-  tsconfig: "tsconfig.build.json",
-  dts: true,
-  sourcemap: true,
-});
